# swagger.yaml
openapi: 3.0.0
info:
  title: Airtime2Cash
  description: In order to convert mobile airtime into cash using the web-based application Airtime2Cash, you must first initiate the process from the platform. Once our admin has confirmed the transaction as successful, 70% of the transferred amount will be credited to the client's wallet, from which the client can withdraw the money to his or her bank account.
  contact:
    name: SQ11B Pod F
    email: airtime2cashpodf@gmail.com
  version: '1.0'
servers:
- url: http://localhost:7000
  description: This is the localhost development server
  variables: {}
paths:
...
  /account/deleteaccount/{Account id}:
    parameters:
      - in: path
        name: Account id
        schema:
          type: string
        required: true
        description: Bank Account Id
    delete:
      tags:
      - User
      summary: Delete Bank Account
      description: Registered bank accounts can be deleted with this endpoint
      operationId: DeleteBankAccount
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...